+++ src/lib.rs
@@ -555,6 +555,7 @@ pub mod io {
     pub mod pdb_io;
     pub mod pqr_io;
     pub mod tpr_io;
+    pub mod traj_cat;
     pub mod traj_io;
     pub mod trr_io;
     mod xdrfile;
@@ -594,7 +595,7 @@ pub mod prelude {
     pub use crate::io::traj_io::{
         FrameData, FrameDataTime, TrajFile, TrajGroupWrite, TrajMasterRead, TrajRangeRead,
         TrajRangeReader, TrajRangeStepReader, TrajRead, TrajReader, TrajStepRead, TrajStepReader,
-        TrajWrite,
+        TrajStepTimeRead, TrajWrite,
     };
     pub use crate::io::trr_io::{TrrGroupWriter, TrrReader, TrrWriter};
     pub use crate::io::xtc_io::{XtcGroupWriter, XtcReader, XtcWriter};